+++ ShopmeWebParent/ShopmeBackEnd/src/main/java/com/shopme/admin/user/export/AbstractExporter.java
@@ -1,4 +1,4 @@
-package com.shopme.admin.user;
+package com.shopme.admin.user.export;
 
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
