+++ ShopmeWebParent/ShopmeBackEnd/src/main/java/com/shopme/admin/user/export/UserCsvExporter.java
@@ -1,5 +1,4 @@
 package com.shopme.admin.user.export;
-import com.shopme.admin.user.AbstractExporter;
 import com.shopme.common.entity.User;
 import org.supercsv.io.CsvBeanWriter;
 import org.supercsv.io.ICsvBeanWriter;
