+++ ShopmeWebParent/ShopmeBackEnd/src/main/java/com/shopme/admin/user/export/UserExcelExporter.java
@@ -1,5 +1,4 @@
 package com.shopme.admin.user.export;
-import com.shopme.admin.user.AbstractExporter;
 import com.shopme.common.entity.User;
 import org.apache.poi.ss.usermodel.CellStyle;
 import org.apache.poi.xssf.usermodel.*;
