+++ ShopmeWebParent/ShopmeBackEnd/src/main/java/com/shopme/admin/user/export/UserPdfExporter.java
@@ -5,7 +5,6 @@ import com.lowagie.text.Font;
 import com.lowagie.text.pdf.PdfPCell;
 import com.lowagie.text.pdf.PdfPTable;
 import com.lowagie.text.pdf.PdfWriter;
-import com.shopme.admin.user.AbstractExporter;
 import com.shopme.common.entity.User;
 
 import javax.servlet.http.HttpServletResponse;
