--- ShopmeWebParent/ShopmeBackEnd/src/main/resources/templates/user_form.html
@@ -1,166 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-<head th:replace="fragments :: page_head(${pageTitle})" />
-
-<body>
-<div class="container-fluid">
-    <div th:replace="navigation :: nav_menu"></div>
-
-    <div>
-        <h2>Manage Users | [[${pageTitle}]]</h2>
-    </div>
-
-    <form th:action="@{/users/save}" method="post" style="max-width: 700px; margin: 0 auto" enctype="multipart/form-data" th:object="${user}" onsubmit="return checkEmailUnique(this)">
-        <input type="hidden" th:field="*{id}" />
-        <div class="border border-secondary rounded p-3">
-            <div class="form-group row">
-                <label class="col-sm-4 col-form-label">E-mail :</label>
-                <div class="col-sm-8">
-                    <input  type="email" class="form-control" th:field="*{email}" required minlength="8" maxlength="128"/>
-                </div>
-            </div>
-
-            <div class="form-group row">
-                <label class="col-sm-4 col-form-label">First Name :</label>
-                <div class="col-sm-8">
-                    <input  type="text" class="form-control" th:field="*{firstName}" required minlength="2" maxlength="45"/>
-                </div>
-            </div>
-
-            <div class="form-group row">
-                <label class="col-sm-4 col-form-label">Last Name :</label>
-                <div class="col-sm-8">
-                    <input  type="text" class="form-control" th:field="*{lastName}" required minlength="2" maxlength="45"/>
-                </div>
-            </div>
-
-            <div class="form-group row">
-                <label class="col-sm-4 col-form-label">Password :</label>
-                <div class="col-sm-8">
-                    <!-- New Account Mode -->
-                    <input  th:if="${user.id == null}" type="password" class="form-control" th:field="*{password}" required minlength="8" maxlength="20"/>
-                    <!-- Edit Mode -->
-                    <input  th:if="${user.id != null}" type="password" class="form-control" th:field="*{password}" minlength="8" maxlength="20"/>
-                </div>
-            </div>
-
-            <div class="form-group row">
-                <label class="col-sm-4 col-form-label">Roles :</label>
-                <div class="col-sm-8">
-                    <th:block th:each="role : ${listRoles}"}>
-                        <input type="checkbox" th:field="*{roles}" th:text="${role.name}" th:value="${role.id}" class="m-2"/>
-                        - <small>[[${role.description}]]</small><br>
-                        </th:block>
-                </div>
-            </div>
-
-            <div class="form-group row">
-                <label class="col-sm-4 col-form-label">Enabled :</label>
-                <div class="col-sm-8">
-                    <input  type="checkbox" th:field="*{enabled}" />
-                </div>
-            </div>
-
-            <!-- User Photo -->
-            <div class="form-group row">
-                <label class="col-sm-4 col-form-label">Photos :</label>
-                <div class="col-sm-8">
-                    <input type="hidden" th:field="*{photos}" />
-                    <input  type="file" id="fileImage" name="image" accept="image/png, image/jpeg" class="mb-2"/>
-                    <img id="thumbnail" alt="Photos Preview" th:src="@{${user.photosImagePath}}" class="img-fluid" style="width: 110px"/>
-                </div>
-            </div>
-
-            <div class="text-center">
-                <input type="submit" value="Save" class="btn btn-primary m-3"/>
-                <input type="button" value="Cancel" class="btn btn-secondary" id="buttonCancel"/>
-            </div>
-        </div>
-    </form>
-
-    <!-- Dialog Box -->
-    <div class="modal fade text-center" id="modalDialog">
-        <div class="modal-dialog">
-            <div class="modal-content">
-                <div class="modal-header">
-                    <h4 class="modal-title" id="modalTitle">Warning</h4>
-                    <button type="button" class="btn-close" data-bs-dismiss="modal"></button>
-                </div>
-
-                <div class="modal-body">
-                    <span id="modalBody"></span>
-                </div>
-
-                <div class="modal-footer">
-                    <button type="button" class="btn btn btn-warning" data-bs-dismiss="modal">Close</button>
-                </div>
-            </div>
-        </div>
-    </div>
-
-    <div th:replace="fragments :: footer"></div>
-</div>
-
-<script type="text/javascript">
-    $(document).ready(function(){
-        $("#buttonCancel").on("click", function(){
-            window.location = "[[@{/users}]]";
-        });
-
-        $("#fileImage").change(function() {
-            fileSize = this.files[0].size;
-            // alert("File size: " + fileSize);
-            if(fileSize > 5242880){
-                this.setCustomValidity("The photo size must be < 10MB");
-                this.reportValidity();
-            }
-            else{
-                this.setCustomValidity("");
-                showImageThumbnail(this);
-            }
-        });
-    });
-
-    function showImageThumbnail(fileInput){
-        var file = fileInput.files[0];
-        var reader = new FileReader();
-        reader.onload = function(e){
-            $("#thumbnail").attr("src", e.target.result);
-        };
-
-        reader.readAsDataURL(file);
-    };
-
-    function checkEmailUnique(form){
-        url = "[[@{/users/check_email}]]";
-        userEmail = $("#email").val();
-        userId = $("#id").val();
-        csrfVal = $("input[name='_csrf']").val();
-        params = {id: userId, email: userEmail, _csrf: csrfVal};
-
-        $.post(url, params, function(response){
-            if (response == "Successful"){
-                form.submit();
-            } else if (response == "Duplicated"){
-                showModalDialog("Warning", "Email address has already been registered : " + userEmail);
-            } else{
-                showModalDialog("Error", "Unknown response from server");
-            }
-        }).fail(function(){
-            showModalDialog("Error", "Could not connect to server");
-        });
-
-        return false;
-    }
-
-    function showModalDialog(title, message){
-        $("#modalTitle").text(title);
-        $("#modalBody").text(message);
-        $("#modalDialog").modal('show');
-    }
-
-</script>
-
-</body>
-
-</html>
