+++ ShopmeWebParent/ShopmeBackEnd/src/main/resources/templates/users/user_form.html
@@ -1,27 +1,22 @@
 <!DOCTYPE html>
 <html lang="en" xmlns:th="http://www.thymeleaf.org">
-<head th:replace="fragments :: page_head('Your Account Details')" />
+<head th:replace="fragments :: page_head(${pageTitle})" />
 
 <body>
 <div class="container-fluid">
     <div th:replace="navigation :: nav_menu"></div>
 
-    <form th:action="@{/account/update}" method="post" style="max-width: 700px; margin: 0 auto" enctype="multipart/form-data" th:object="${user}">
+    <div>
+        <h2>Manage Users | [[${pageTitle}]]</h2>
+    </div>
+
+    <form th:action="@{/users/save}" method="post" style="max-width: 700px; margin: 0 auto" enctype="multipart/form-data" th:object="${user}" onsubmit="return checkEmailUnique(this)">
         <input type="hidden" th:field="*{id}" />
-
-        <div class="text-center">
-            <h2>Your Account Details</h2>
-        </div>
-
-        <div th:if="${message != null}" class="alert alert-success text-center">
-            [[${message}]]
-        </div>
-
         <div class="border border-secondary rounded p-3">
             <div class="form-group row">
                 <label class="col-sm-4 col-form-label">E-mail :</label>
                 <div class="col-sm-8">
-                    <input  type="email" class="form-control" th:field="*{email}" readonly="readonly"/>
+                    <input  type="email" class="form-control" th:field="*{email}" required minlength="8" maxlength="128"/>
                 </div>
             </div>
 
@@ -39,28 +34,30 @@
                 </div>
             </div>
 
-            <!-- Password -->
             <div class="form-group row">
                 <label class="col-sm-4 col-form-label">Password :</label>
                 <div class="col-sm-8">
+                    <!-- New Account Mode -->
+                    <input  th:if="${user.id == null}" type="password" class="form-control" th:field="*{password}" required minlength="8" maxlength="20"/>
                     <!-- Edit Mode -->
-                    <input  type="password" class="form-control" th:field="*{password}" minlength="8" maxlength="20" oninput="checkPasswordMatch(document.getElementById('confirmPassword'))"/>
-                </div>
-            </div>
-
-            <!-- Confirm Password -->
-            <div class="form-group row">
-                <label class="col-sm-4 col-form-label">Confirm Password :</label>
-                <div class="col-sm-8">
-                    <!-- Edit Mode -->
-                    <input  type="password" id="confirmPassword" class="form-control" minlength="8" maxlength="20" oninput="checkPasswordMatch(this)"/>
+                    <input  th:if="${user.id != null}" type="password" class="form-control" th:field="*{password}" minlength="8" maxlength="20"/>
                 </div>
             </div>
 
             <div class="form-group row">
                 <label class="col-sm-4 col-form-label">Roles :</label>
                 <div class="col-sm-8">
-                    <span><b>[[${user.roles}]]</b></span>
+                    <th:block th:each="role : ${listRoles}"}>
+                        <input type="checkbox" th:field="*{roles}" th:text="${role.name}" th:value="${role.id}" class="m-2"/>
+                        - <small>[[${role.description}]]</small><br>
+                        </th:block>
+                </div>
+            </div>
+
+            <div class="form-group row">
+                <label class="col-sm-4 col-form-label">Enabled :</label>
+                <div class="col-sm-8">
+                    <input  type="checkbox" th:field="*{enabled}" />
                 </div>
             </div>
 
@@ -105,24 +102,7 @@
 </div>
 
 <script type="text/javascript">
-    $(document).ready(function(){
-        $("#buttonCancel").on("click", function(){
-            window.location = "[[@{/users}]]";
-        });
-
-        $("#fileImage").change(function() {
-            fileSize = this.files[0].size;
-            // alert("File size: " + fileSize);
-            if(fileSize > 5242880){
-                this.setCustomValidity("The photo size must be < 10MB");
-                this.reportValidity();
-            }
-            else{
-                this.setCustomValidity("");
-                showImageThumbnail(this);
-            }
-        });
-    });
+    moduleURL = "[[@{/users}]]";
 
     function showImageThumbnail(fileInput){
         var file = fileInput.files[0];
@@ -134,15 +114,37 @@
         reader.readAsDataURL(file);
     }
 
-    function checkPasswordMatch(confirmPassword){
-        if(confirmPassword.value != $("#password").val()){
-            confirmPassword.setCustomValidity("Passwords do not mathch");
-        } else {
-            confirmPassword.setCustomValidity("");
-        }
+    function checkEmailUnique(form){
+        url = "[[@{/users/check_email}]]";
+        userEmail = $("#email").val();
+        userId = $("#id").val();
+        csrfVal = $("input[name='_csrf']").val();
+        params = {id: userId, email: userEmail, _csrf: csrfVal};
+
+        $.post(url, params, function(response){
+            if (response == "Successful"){
+                form.submit();
+            } else if (response == "Duplicated"){
+                showModalDialog("Warning", "Email address has already been registered : " + userEmail);
+            } else{
+                showModalDialog("Error", "Unknown response from server");
+            }
+        }).fail(function(){
+            showModalDialog("Error", "Could not connect to server");
+        });
+
+        return false;
+    }
+
+    function showModalDialog(title, message){
+        $("#modalTitle").text(title);
+        $("#modalBody").text(message);
+        $("#modalDialog").modal('show');
     }
 
 </script>
+
+<script th:src="@{/js/common_form.js}"></script>
 
 </body>
